data_processing/requirements.txt
requests
confluent-kafka
google-api-python-client
beautifulsoup4
python-dotenv
nltk